fix a big bug

+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/user_services_functions/start_user_services.go
@@ -83,6 +83,9 @@ func StartUserServices(
 		userServiceFilters := &service.ServiceFilters{
 			IDs: serviceIDsToRemove,
 		}
+		if len(serviceIDsToRemove) == 0 {
+			return
+		}
 		_, failedToDestroyGUIDs, err := destroyUserServicesUnlocked(ctx, enclaveID, userServiceFilters, serviceRegistrations, enclaveFreeIpProviders, dockerManager)
 		if err != nil {
 			for serviceID, _ := range serviceIDsToRemove {
